Ok, Finally got it working. now I just need to do everything else in the world to do on this thing lol

+++ Crimson Engine/GUI_Element_Size.h
@@ -4,9 +4,9 @@
 struct GUI_Element_Size : public GUI_Element
 {
 	ElementType TypeOfElement = Size;
-	uint8_t Width;
-	uint8_t Height;
-	GUI_Element_Size(uint8_t W, uint8_t H) {
+	int Width;
+	int Height;
+	GUI_Element_Size(int W, int H) {
 		Width = W;
 		Height = H;
 	};
